fix: cleanup example file
