+++ package.json
@@ -13,6 +13,7 @@
   "scripts": {
     "build": "rm -rf dist && tsc -p .",
     "test": "vitest",
+    "dev": "esno ./playground/usage-example.ts",
     "coverage": "vitest run --coverage"
   },
   "author": "Valentin Lica",
@@ -25,12 +26,13 @@
     "@types/cheerio": "^0.22.31",
     "@types/node-fetch": "^2.6.2",
     "cheerio": "^1.0.0-rc.12",
-    "node-fetch": "^2.6.7"
+    "node-fetch": "^3.2.10"
   },
   "devDependencies": {
+    "esno": "*",
+    "picocolors": "^1.0.0",
     "typescript": "^4.7.4",
-    "vite": "^3.0.5",
-    "vitest": "^0.19.0",
-    "picocolors": "^1.0.0"
+    "vite": "^3.0.7",
+    "vitest": "^0.21.1"
   }
 }
