+++ tsconfig.json
@@ -12,6 +12,9 @@
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true
   },
+  "types": [
+    "vitest/globals"
+  ],
   "include": [
     "src"
   ]
